+++ tests/integration/index.js
@@ -37,7 +37,7 @@ const publicationDeleteTests = require('./publication/publication-delete.test')
 const publicationPostTests = require('./publication/publication-post.test')
 const publicationPatchTests = require('./publication/publication-patch.test')
 
-const readerCreateTests = require('./reader-create.test')
+const readerCreateTests = require('./reader/reader-post.test')
 const readerGetTests = require('./reader/reader-get.test')
 const readerGetTestsOld = require('./deprecated/reader-get-old.test') // deprecated
