+++ tests/integration/reader/reader-post.test.js
@@ -1,6 +1,6 @@
 const request = require('supertest')
 const tap = require('tap')
-const { getToken, destroyDB } = require('../utils/testUtils')
+const { getToken, destroyDB } = require('../../utils/testUtils')
 
 const test = async app => {
   const token = getToken()
